Large scale modification

+++ datastore/rdbRoomUserStore.go
@@ -24,7 +24,7 @@ func rdbCreateRoomUserStore(db string) {
 	}
 }
 
-func rdbDeleteAndInsertRoomUsers(db string, roomUsers []*model.RoomUser) error {
+func rdbInsertRoomUsers(db string, roomUsers []*model.RoomUser, opts ...InsertRoomUsersOption) error {
 	master := RdbStore(db).master()
 	trans, err := master.Begin()
 	if err != nil {
@@ -33,14 +33,21 @@ func rdbDeleteAndInsertRoomUsers(db string, roomUsers []*model.RoomUser) error {
 		return err
 	}
 
-	query := fmt.Sprintf("DELETE FROM %s WHERE room_id=:roomId;", tableNameRoomUser)
-	params := map[string]interface{}{"roomId": roomUsers[0].RoomID}
-	_, err = trans.Exec(query, params)
-	if err != nil {
-		trans.Rollback()
-		err := errors.Wrap(err, "An error occurred while recreating roomUser")
-		logger.Error(err.Error())
-		return err
+	opt := insertRoomUsersOptions{}
+	for _, o := range opts {
+		o(&opt)
+	}
+
+	if opt.beforeClean {
+		query := fmt.Sprintf("DELETE FROM %s WHERE room_id=:roomId;", tableNameRoomUser)
+		params := map[string]interface{}{"roomId": roomUsers[0].RoomID}
+		_, err = trans.Exec(query, params)
+		if err != nil {
+			trans.Rollback()
+			err := errors.Wrap(err, "An error occurred while recreating roomUser")
+			logger.Error(err.Error())
+			return err
+		}
 	}
 
 	for _, roomUser := range roomUsers {
@@ -64,43 +71,52 @@ func rdbDeleteAndInsertRoomUsers(db string, roomUsers []*model.RoomUser) error {
 	return nil
 }
 
-func rdbInsertRoomUsers(db string, roomUsers []*model.RoomUser) error {
-	master := RdbStore(db).master()
-	trans, err := master.Begin()
+func rdbSelectRoomUsers(db string, opts ...SelectRoomUsersOption) ([]*model.RoomUser, error) {
+	replica := RdbStore(db).replica()
+
+	opt := selectRoomUsersOptions{}
+	for _, o := range opts {
+		o(&opt)
+	}
+
+	var roomUsers []*model.RoomUser
+	var userIDsQuery string
+	var userIDsParams map[string]interface{}
+	var roomIDParams map[string]interface{}
+
+	if opt.roomID != "" {
+		roomIDParams = map[string]interface{}{"roomId": opt.roomID}
+	}
+	if opt.userIDs != nil {
+		userIDsQuery, userIDsParams = utils.MakePrepareForInExpression(opt.userIDs)
+	}
+	params := make(map[string]interface{}, len(userIDsParams)+len(roomIDParams))
+	params = utils.MergeMap(params, userIDsParams, roomIDParams)
+
+	query := fmt.Sprintf("SELECT * FROM %s WHERE", tableNameRoomUser)
+	if opt.roomID != "" {
+		query = fmt.Sprintf("%s room_id=:roomId", query)
+	}
+	if opt.roomID != "" && opt.userIDs != nil {
+		query = fmt.Sprintf("%s AND ", query)
+	}
+	if opt.userIDs != nil {
+		query = fmt.Sprintf("%s user_id IN (%s)", query, userIDsQuery)
+	}
+
+	var err error
+	if params == nil {
+		_, err = replica.Select(&roomUsers, query)
+	} else {
+		_, err = replica.Select(&roomUsers, query, params)
+	}
 	if err != nil {
-		err := errors.Wrap(err, "An error occurred while inserting roomUsers")
+		err := errors.Wrap(err, "An error occurred while getting room users")
 		logger.Error(err.Error())
-		return err
+		return nil, err
 	}
 
-	for _, roomUser := range roomUsers {
-		ru, err := rdbSelectRoomUser(db, roomUser.RoomID, roomUser.UserID)
-		if err != nil {
-			trans.Rollback()
-			err := errors.Wrap(err, "An error occurred while inserting roomUsers")
-			logger.Error(err.Error())
-			return err
-		}
-		if ru == nil {
-			err = trans.Insert(roomUser)
-			if err != nil {
-				trans.Rollback()
-				err := errors.Wrap(err, "An error occurred while inserting roomUsers")
-				logger.Error(err.Error())
-				return err
-			}
-		}
-	}
-
-	err = trans.Commit()
-	if err != nil {
-		trans.Rollback()
-		err := errors.Wrap(err, "An error occurred while inserting roomUsers")
-		logger.Error(err.Error())
-		return err
-	}
-
-	return nil
+	return roomUsers, nil
 }
 
 func rdbSelectRoomUser(db, roomID, userID string) (*model.RoomUser, error) {
@@ -153,46 +169,10 @@ WHERE room_id IN (
 	return nil, nil
 }
 
-func rdbSelectRoomUsersByRoomID(db, roomID string) ([]*model.RoomUser, error) {
+func rdbSelectUserIDsOfRoomUser(db string, roomID string, opts ...SelectUserIDsOfRoomUserOption) ([]string, error) {
 	replica := RdbStore(db).replica()
 
-	var roomUsers []*model.RoomUser
-	query := fmt.Sprintf("SELECT room_id, user_id, unread_count FROM %s WHERE room_id=:roomId;", tableNameRoomUser)
-	params := map[string]interface{}{
-		"roomId": roomID,
-	}
-	_, err := replica.Select(&roomUsers, query, params)
-	if err != nil {
-		err := errors.Wrap(err, "An error occurred while getting roomUsers by roomId")
-		logger.Error(err.Error())
-		return nil, err
-	}
-
-	return roomUsers, nil
-}
-
-func rdbSelectRoomUsersByUserID(db, userID string) ([]*model.RoomUser, error) {
-	replica := RdbStore(db).replica()
-
-	var roomUsers []*model.RoomUser
-	query := fmt.Sprintf("SELECT * FROM %s WHERE user_id=:userId;", tableNameRoomUser)
-	params := map[string]interface{}{
-		"userId": userID,
-	}
-	_, err := replica.Select(&roomUsers, query, params)
-	if err != nil {
-		err := errors.Wrap(err, "An error occurred while getting roomUsers by userId")
-		logger.Error(err.Error())
-		return nil, err
-	}
-
-	return roomUsers, nil
-}
-
-func rdbSelectUserIDsOfRoomUser(db string, roomID string, opts ...RoomUserOption) ([]string, error) {
-	replica := RdbStore(db).replica()
-
-	opt := roomUserOptions{}
+	opt := selectUserIDsOfRoomUserOptions{}
 	for _, o := range opts {
 		o(&opt)
 	}
@@ -223,42 +203,7 @@ func rdbSelectUserIDsOfRoomUser(db string, roomID string, opts ...RoomUserOption
 	return userIDs, nil
 }
 
-func rdbSelectRoomUsersByRoomIDAndUserIDs(db string, roomID *string, userIDs []string) ([]*model.RoomUser, error) {
-	replica := RdbStore(db).replica()
-
-	var roomUsers []*model.RoomUser
-	var userIDsQuery string
-	var userIDsParams map[string]interface{}
-	var roomIDParams map[string]interface{}
-	if userIDs != nil {
-		userIDsQuery, userIDsParams = utils.MakePrepareForInExpression(userIDs)
-	}
-	if roomID != nil {
-		roomIDParams = map[string]interface{}{"roomId": roomID}
-	}
-	params := utils.MergeMap(userIDsParams, roomIDParams)
-
-	query := fmt.Sprintf("SELECT * FROM %s WHERE ", tableNameRoomUser)
-	if roomID != nil {
-		query = fmt.Sprintf("%s room_id=:roomId", query)
-	}
-	if roomID != nil && userIDs != nil {
-		query = fmt.Sprintf("%s AND ", query)
-	}
-	if userIDs != nil {
-		query = fmt.Sprintf("%s user_id IN (%s)", query, userIDsQuery)
-	}
-	_, err := replica.Select(&roomUsers, query, params)
-	if err != nil {
-		err := errors.Wrap(err, "An error occurred while getting room users")
-		logger.Error(err.Error())
-		return nil, err
-	}
-
-	return roomUsers, nil
-}
-
-func rdbUpdateRoomUser(db string, ru *model.RoomUser) (*model.RoomUser, error) {
+func rdbUpdateRoomUser(db string, ru *model.RoomUser) error {
 	master := RdbStore(db).master()
 
 	query := fmt.Sprintf("UPDATE %s SET unread_count=:unreadCount WHERE room_id=:roomId AND user_id=:userId;", tableNameRoomUser)
@@ -271,70 +216,13 @@ func rdbUpdateRoomUser(db string, ru *model.RoomUser) (*model.RoomUser, error) {
 	if err != nil {
 		err := errors.Wrap(err, "An error occurred while updating room user")
 		logger.Error(err.Error())
-		return nil, err
+		return err
 	}
 
-	return ru, nil
+	return nil
 }
 
-// func rdbUpdateRoomUser(db string, roomUser *model.RoomUser) (*model.RoomUser, error) {
-// 	master := RdbStore(db).master()
-// 	trans, err := master.Begin()
-// 	if err != nil {
-// 		return nil, errors.Wrap(err, "An error occurred while transaction beginning")
-// 	}
-
-// 	updateQuery := ""
-// 	params := map[string]interface{}{
-// 		"roomId": roomUser.RoomID,
-// 		"userId": roomUser.UserID,
-// 	}
-// 	if roomUser.UnreadCount != nil {
-// 		params["unreadCount"] = roomUser.UnreadCount
-// 		updateQuery = "unread_count=:unreadCount"
-// 	}
-// 	// if roomUser.MetaData != nil {
-// 	// 	params["metaData"] = roomUser.MetaData
-// 	// 	if updateQuery == "" {
-// 	// 		updateQuery = "meta_data=:metaData"
-// 	// 	} else {
-// 	// 		updateQuery = utils.AppendStrings(updateQuery, ",", "meta_data=:metaData")
-// 	// 	}
-// 	// }
-// 	if updateQuery != "" {
-// 		query := utils.AppendStrings("UPDATE ", tableNameRoomUser, " SET "+updateQuery+" WHERE room_id=:roomId AND user_id=:userId;")
-// 		_, err = trans.Exec(query, params)
-// 		if err != nil {
-// 			trans.Rollback()
-// 			return nil, errors.Wrap(err, "An error occurred while updating room's users")
-// 		}
-
-// 		if roomUser.UnreadCount != nil {
-// 			query = utils.AppendStrings("UPDATE ", tableNameUser,
-// 				" SET unread_count=(SELECT SUM(unread_count) FROM ", tableNameRoomUser,
-// 				" WHERE user_id=:userId1) WHERE user_id=:userId2;")
-// 			params = map[string]interface{}{
-// 				"userId1": roomUser.UserID,
-// 				"userId2": roomUser.UserID,
-// 			}
-// 			_, err = trans.Exec(query, params)
-// 			if err != nil {
-// 				trans.Rollback()
-// 				return nil, errors.Wrap(err, "An error occurred while updating user unread count")
-// 			}
-// 		}
-// 	}
-
-// 	err = trans.Commit()
-// 	if err != nil {
-// 		trans.Rollback()
-// 		return nil, errors.New("An error occurred while commit updating room's user")
-// 	}
-
-// 	return roomUser, nil
-// }
-
-func rdbDeleteRoomUser(db, roomID string, userIDs []string) error {
+func rdbDeleteRoomUsers(db, roomID string, userIDs []string) error {
 	master := RdbStore(db).master()
 	trans, err := master.Begin()
 	if err != nil {
@@ -402,3 +290,60 @@ func rdbDeleteRoomUser(db, roomID string, userIDs []string) error {
 
 	return nil
 }
+
+// func rdbUpdateRoomUser(db string, roomUser *model.RoomUser) (*model.RoomUser, error) {
+// 	master := RdbStore(db).master()
+// 	trans, err := master.Begin()
+// 	if err != nil {
+// 		return nil, errors.Wrap(err, "An error occurred while transaction beginning")
+// 	}
+
+// 	updateQuery := ""
+// 	params := map[string]interface{}{
+// 		"roomId": roomUser.RoomID,
+// 		"userId": roomUser.UserID,
+// 	}
+// 	if roomUser.UnreadCount != nil {
+// 		params["unreadCount"] = roomUser.UnreadCount
+// 		updateQuery = "unread_count=:unreadCount"
+// 	}
+// 	// if roomUser.MetaData != nil {
+// 	// 	params["metaData"] = roomUser.MetaData
+// 	// 	if updateQuery == "" {
+// 	// 		updateQuery = "meta_data=:metaData"
+// 	// 	} else {
+// 	// 		updateQuery = utils.AppendStrings(updateQuery, ",", "meta_data=:metaData")
+// 	// 	}
+// 	// }
+// 	if updateQuery != "" {
+// 		query := utils.AppendStrings("UPDATE ", tableNameRoomUser, " SET "+updateQuery+" WHERE room_id=:roomId AND user_id=:userId;")
+// 		_, err = trans.Exec(query, params)
+// 		if err != nil {
+// 			trans.Rollback()
+// 			return nil, errors.Wrap(err, "An error occurred while updating room's users")
+// 		}
+
+// 		if roomUser.UnreadCount != nil {
+// 			query = utils.AppendStrings("UPDATE ", tableNameUser,
+// 				" SET unread_count=(SELECT SUM(unread_count) FROM ", tableNameRoomUser,
+// 				" WHERE user_id=:userId1) WHERE user_id=:userId2;")
+// 			params = map[string]interface{}{
+// 				"userId1": roomUser.UserID,
+// 				"userId2": roomUser.UserID,
+// 			}
+// 			_, err = trans.Exec(query, params)
+// 			if err != nil {
+// 				trans.Rollback()
+// 				return nil, errors.Wrap(err, "An error occurred while updating user unread count")
+// 			}
+// 		}
+// 	}
+
+// 	err = trans.Commit()
+// 	if err != nil {
+// 		trans.Rollback()
+// 		return nil, errors.New("An error occurred while commit updating room's user")
+// 	}
+
+// 	return roomUser, nil
+// }
